Flatten commands.ts
